perf(frontend): lazy load the unified search modal
